Simplify the logic to only send POST events to the kaleidoscope as a result of user interaction. Instead of using multiple states (settingUp, canFetchSettings, wasItUs) take advantage of the simplifying logic that the only time we _don't_ want to send POST events is when we are updating the UI as a result of fetching updated settings from the kaleidoscope.
Also add logic to cancel all pending fetch requests so that they don't come in _after_ we've made a
change in the UI and override that change.

Update the throttling logic so that we only throttle dragging the slider (the only way we can really
generate too many POST events). This ensures the POST request from when we stop dragging the slider
in onStopTrackingTouch will always go through. Otherwise, that last change may get throttled and lost.