Switch to a new makePost implementation that uses the okhttp3 library with it's existing client queue. This is a nice optimization as it already implements okhttp3 threaded queuing so we don't need to create a new thread for every request.
The key to getting this to work properly was to create the request using getBytes
so that okhttp3 doesn't append the character encoding. Without this workaround,
the character encoding caused a 500 error as the Kaleidoscope doesn't handle encoding
strings properly. See https://github.com/square/okhttp/issues/2099 for details on this
workaround.

Fix some leaked connections by properly ensuring we always use/close the response
body in our onResponse methods.

Fix some random crashes when canceling requests in the okhttp3 queue by ensuring
there is a tag object before attempting to compare it to the requested tag.